add create section color menu
